backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
transformers>=4.46.0
torch>=2.0.0
torchaudio>=2.0.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
llama-cpp-python>=0.2.0
ffmpeg-python==0.2.0
pillow==10.4.0
pydantic==2.9.2
huggingface-hub>=0.25.0
aiofiles==24.1.0
requests>=2.31.0
